Callback to allow ImageSource to be updated by WSGI middleware
Adds postprocess_image method to the base Server that allows WSGI
middleware to post process the ImageSource prior to it being returned in
the response. Commonly used for applying a graphic watermark or applying
a filter to the image.

The callback is associated with the key mapproxy.postprocess in the WSGI
environment. It is passed the ImageSource and is expected to return an
ImageSource.

The postprocess_image method is called from service/wms and service/tile
prior to the ImageSource being returned in a response.

Added unit tests for postprocess_image method.