use dummy access_contraints to clarify license
The access_contraints option allows to document the license of the data
that MapProxy serves. MapProxy itself does not provide any data and it's
up to the user to add free or closed data sources.

Some example and test configurations mentioned the Creative Commons
license to show how a typical access_contraints configuration can look like.

This commit removes all references to Creative Commons to make it clear that
MapProxy and all the example and test files are _not_ licensed as
Creative Commons.