dynamically calculate required QUADs for MESH transformation
The number of QUADs is calculated dynamically to keep the deviation in
the image transformation below one pixel. Image transformations for large
map scales can be transformed with 1-4 QUADs most of the time. For low scales,
transform_meshes can generate a few hundred QUADs.

A fixed set of 8x8 QUADs was used before this patch.

Closes #285