capabilities: Cache the capabilities document as static per config
Each time the config file gets loaded (or reloaded), all server objects
are recreated. So it is safe do store the capabilities document, once
generated, as a static document. The first request to getCapabilities
will render de document and store it as static. All other requests will
get the static document until the configuration is reloaded, when a new
document will be rendered and stored again.
The Cache-Status non-standard HTTP header will inform if the capabilities document
was rendered or was retrieved from cache (static variable).