capabilities: Cache the capabilities document as static per config+request+environment
Each time the config file gets loaded (or reloaded), all server objects
are recreated. So it should be safe to store the capabilities document, once
generated, as a static document if we generate one document for different
request version and/or mime_type, and also consider the mapproxy.authorization .
The first request to getCapabilities will render de document and store it as static.
All other requests will get the static document until the configuration is reloaded,
there are diferent values in version or mime_type, differente environment variables
like HTTP_HOST, or the request is using mapproxy.authorize, when a new
document will be rendered and stored again.
The Cache-Status non-standard HTTP header will inform if the capabilities document
was rendered or was retrieved from cache (static variable).