fix capabilities documents and extend demo page
Fixes #495 and extends the functionality of the demo page. The user can now choose to view the capabilities documents either for the internal address (e.g. 10.0.13.37) or the external address (e.g. example.org). This makes debugging easier when the service is running behind a reverse proxy and throw an "internal error" as before.