register_service_configuration(): add an optional schema_service parameter
Since the introduction of this method, JSON Schema validation was added.
When a plugin adds a new service which requires configuration elements in
the .yaml configuration file, it is also needed to extend the JSON schema
to avoid warnings at startup. Will be used by https://github.com/rouault/mapproxy_hips.